Refactor code modify Admin page
