+++ src/components/handshakes/betting/Feed/Filter/Filter.jsx
@@ -7,7 +7,8 @@ import { BigNumber } from 'bignumber.js';
 import local from '@/services/localStore';
 import { APP, API_URL } from '@/constants';
 import { loadMatches, loadHandshakes, checkFreeAvailable } from '@/reducers/betting/action';
-import { SIDE } from '@/components/handshakes/betting/Feed/BetHandshakeHandler';
+import { SIDE } from '@/components/handshakes/betting/constants.js';
+
 import { getBalance, parseBigNumber } from '@/components/handshakes/betting/utils';
 import GA from '@/services/googleAnalytics';
 // components
