+++ src/components/handshakes/betting/Feed/ShakeFree/ShakeFree.jsx
@@ -20,7 +20,8 @@ import { validateBet } from '@/components/handshakes/betting/validation.js';
 import { MESSAGE } from '@/components/handshakes/betting/message.js';
 
 
-import { BetHandshakeHandler, SIDE } from '@/components/handshakes/betting/Feed/BetHandshakeHandler';
+import { BetHandshakeHandler } from '@/components/handshakes/betting/Feed/BetHandshakeHandler';
+import { SIDE } from '@/components/handshakes/betting/constants.js';
 
 import './ShakeFree.scss';
 import Toggle from './../Toggle';
